+++ Pandora/Paradigm/Inventory/Ability/Modifiable.hs
@@ -6,4 +6,4 @@ import Pandora.Paradigm.Inventory.Ability.Replaceable (Replaceable)
 
 class (Viewable i, Replaceable i) => Modifiable i where
 	type Modification i e r :: *
-	modify_ :: Modification i e r
+	modify :: Modification i e r